# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для системы электронной очереди",
        "title": "Electronic Queue API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/admin/ads": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список всех рекламных материалов без самих изображений.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Получить список всех рекламных материалов (Админ)",
                "responses": {
                    "200": {
                        "description": "Список рекламных материалов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.AdResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Загружает новое рекламное объявление.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Создать рекламный материал (Админ)",
                "parameters": [
                    {
                        "description": "Данные для создания",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateAdRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Созданный материал",
                        "schema": {
                            "$ref": "#/definitions/models.AdResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка в запросе",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/ads/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает полную информацию о рекламном материале, включая изображение.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Получить рекламный материал по ID (Админ)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламного материала",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Рекламный материал",
                        "schema": {
                            "$ref": "#/definitions/models.AdResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет рекламное объявление по ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Удалить рекламный материал (Админ)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламного материала",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Удалено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка удаления",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновляет существующее рекламное объявление.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Обновить рекламный материал (Админ)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID рекламного материала",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateAdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный материал",
                        "schema": {
                            "$ref": "#/definitions/models.AdResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка в запросе",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/create/administrator": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает нового пользователя с ролью \"администратор\". Требует INTERNAL_API_KEY.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Создать нового администратора (Админ)",
                "parameters": [
                    {
                        "description": "Данные нового администратора",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateAdministratorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Администратор успешно создан",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка: неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Ошибка: логин уже занят",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/create/doctor": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает нового пользователя с ролью \"врач\". Требует INTERNAL_API_KEY.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Создать нового врача (Админ)",
                "parameters": [
                    {
                        "description": "Данные нового врача",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateDoctorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Врач успешно создан",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка: неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Ошибка: логин уже занят",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/create/registrar": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает нового пользователя с ролью \"регистратор\". Требует INTERNAL_API_KEY.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Создать нового регистратора (Админ)",
                "parameters": [
                    {
                        "description": "Данные нового регистратора",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateRegistrarRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Регистратор успешно создан",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка: неверный запрос",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Ошибка: логин уже занят",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/processes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список всех бизнес-процессов и их текущее состояние (включен/отключен).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Получить статусы всех бизнес-процессов (Админ)",
                "responses": {
                    "200": {
                        "description": "Список процессов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BusinessProcess"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/processes/{name}": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Включает или отключает указанный бизнес-процесс.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Обновить статус бизнес-процесса (Админ)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя процесса",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новое состояние",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateProcessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Обновленный процесс",
                        "schema": {
                            "$ref": "#/definitions/models.BusinessProcess"
                        }
                    },
                    "400": {
                        "description": "Ошибка в запросе",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Процесс не найден",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.UpdateProcessRequest": {
            "type": "object",
            "properties": {
                "is_enabled": {
                    "type": "boolean"
                }
            }
        },
        "models.AdResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "duration_sec": {
                    "description": "omitempty уберет поле, если оно nil",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "media_type": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "reception_on": {
                    "type": "boolean"
                },
                "repeat_count": {
                    "description": "omitempty уберет поле, если оно nil",
                    "type": "integer"
                },
                "schedule_on": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateAdRequest": {
            "type": "object",
            "properties": {
                "duration_sec": {
                    "description": "--- ИЗМЕНЕНИЕ: gt=0 на gte=1 ---",
                    "type": "integer",
                    "minimum": 1
                },
                "is_enabled": {
                    "type": "boolean"
                },
                "picture": {
                    "type": "string"
                },
                "reception_on": {
                    "type": "boolean"
                },
                "repeat_count": {
                    "description": "--- ИЗМЕНЕНИЕ: gt=0 на gte=1 ---",
                    "type": "integer",
                    "minimum": 1
                },
                "schedule_on": {
                    "type": "boolean"
                },
                "video": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        }
    }
}